requirements.txt
h5py==2.7.0
scipy==1.1.0
numpy==1.13.3
matplotlib==2.1.0
typing==3.6.2
OpenEXR==1.3.2
Pillow==5.4.1
bpy==1.2.3
mathutils==2.79.2
nibabel==2.3.3
open3d==0.1.1
pptk==0.1.0
pykitti==0.3.0
scikit_learn==0.20.2
tensorflow==1.12.0
tensorpack==0.9.0.1
termcolor==1.1.0
torch==1.0.1
torchvision==0.2.1
